Filters work but you can't undo a filter
